Move logic to lib.rs
