+++ src/lib.rs
@@ -1,3 +1,41 @@
+extern crate futures;
+extern crate hyper;
+extern crate hyper_tls;
+extern crate tokio_core;
+
+use std::io::{self, Write};
+use futures::{Future, Stream};
+use hyper::Client;
+use hyper_tls::HttpsConnector;
+use tokio_core::reactor::Core;
+
+pub struct Currencies {
+}
+
+impl Currencies {
+    pub fn get(&self) {
+        let mut core = Core::new().unwrap();
+        let handle = core.handle();
+        let client = Client::configure()
+            .connector(HttpsConnector::new(4, &handle).unwrap())
+            .build(&handle);
+
+        let uri = "https://api.zaif.jp/api/1/currencies/btc".parse().unwrap();
+
+        let work = client.get(uri).and_then(|res| {
+            println!("Response: {}", res.status());
+
+            res.body().for_each(|chunk| {
+                io::stdout()
+                    .write_all(&chunk)
+                    .map(|_| ())
+                    .map_err(From::from)
+            })
+        });
+        core.run(work).unwrap();
+    }
+}
+
 #[cfg(test)]
 mod tests {
     #[test]
