+++ src/main.rs
@@ -1,32 +1,8 @@
-extern crate futures;
-extern crate hyper;
-extern crate hyper_tls;
-extern crate tokio_core;
+extern crate zaif_api;
 
-use std::io::{self, Write};
-use futures::{Future, Stream};
-use hyper::Client;
-use hyper_tls::HttpsConnector;
-use tokio_core::reactor::Core;
+use zaif_api::Currencies;
 
 fn main() {
-    let mut core = Core::new().unwrap();
-    let handle = core.handle();
-    let client = Client::configure()
-        .connector(HttpsConnector::new(4, &handle).unwrap())
-        .build(&handle);
-
-    let uri = "https://api.zaif.jp/api/1/currencies/btc".parse().unwrap();
-
-    let work = client.get(uri).and_then(|res| {
-        println!("Response: {}", res.status());
-
-        res.body().for_each(|chunk| {
-            io::stdout()
-                .write_all(&chunk)
-                .map(|_| ())
-                .map_err(From::from)
-        })
-    });
-    core.run(work).unwrap();
+    let api = Currencies{};
+    api.get();
 }
